requirements.txt
paho-mqtt==2.1.0
aiorwlock==1.5.0
tortoise-orm==0.25.1
bleak==1.0.1
langchain==0.3.26
langchain_community==0.3.27
langchain_openai==0.3.28
langchain_perplexity==0.1.1
langchain-ollama==0.3.6
fastapi[standard]==0.111.0
matplotlib==3.9.2
